Modified initial sync script to handle genesis block correctly.

+++ scripts/sync_coind_initial.php
@@ -2,7 +2,7 @@
 $host_not_required = TRUE;
 include(realpath(dirname(__FILE__))."/../includes/connect.php");
 
-if ($argv) $_REQUEST['key'] = $argv[1];
+if (!empty($argv)) $_REQUEST['key'] = $argv[1];
 
 if ($_REQUEST['key'] == $GLOBALS['cron_key_string']) {
 	$start_time = microtime(true);
@@ -43,36 +43,21 @@ if ($_REQUEST['key'] == $GLOBALS['cron_key_string']) {
 		else die("Error, that block was not found (".$r->rowCount().").");
 	}
 	else {
-		$game->delete_reset_game('reset');
+		$r = $app->run_query("SELECT * FROM blocks WHERE block_id='0' AND game_id='".$game->db_game['game_id']."';");
+		
+		if ($r->rowCount() == 0) {
+			$game->delete_reset_game('reset');
 
-		$q = "DELETE FROM addresses WHERE game_id='".$game->db_game['game_id']."';";
-		$app->run_query($q);
-
-		$genesis_hash = $coin_rpc->getblockhash(0);
-		echo "genesis hash: ".$genesis_hash."<br/>\n";
-
-		$current_hash = $genesis_hash;
-
-		$blocks[$block_height] = new block($coin_rpc->getblock($current_hash), $block_height, $current_hash);
-		$tx_hash = $blocks[$block_height]->json_obj['tx'][0];
-		$transactions[0] = new transaction($tx_hash, "", false, $block_height);
-	
-		$output_address = $game->create_or_fetch_address("genesis_address", true, false, false, false);
-	
-		$q = "INSERT INTO transactions SET game_id='".$game->db_game['game_id']."', amount='".$game->db_game['pow_reward']."', transaction_desc='coinbase', tx_hash='".$tx_hash."', address_id=".$output_address['address_id'].", block_id='".$block_height."', time_created='".time()."';";
-		$r = $app->run_query($q);
-		$transaction_id = $app->last_insert_id();
-	
-		$q = "INSERT INTO transaction_ios SET spend_status='unspent', instantly_mature=0, game_id='".$game->db_game['game_id']."', user_id=NULL, address_id='".$output_address['address_id']."'";
-		$q .= ", create_transaction_id='".$transaction_id."', amount='".$game->db_game['pow_reward']."', create_block_id='".$block_height."';";
-		$r = $app->run_query($q);
-	
-		$q = "INSERT INTO blocks SET game_id='".$game->db_game['game_id']."', block_hash='".$genesis_hash."', block_id='".$block_height."', time_created='".time()."';";
-		$r = $app->run_query($q);
-	
-		echo "Added the genesis transaction!<br/>\n";
-	
-		$current_hash = $blocks[$block_height]->json_obj['nextblockhash'];
+			$q = "DELETE FROM addresses WHERE game_id='".$game->db_game['game_id']."';";
+			$app->run_query($q);
+		
+			$returnvals = $game->add_genesis_block($coin_rpc);
+			$current_hash = $returnvals['nextblockhash'];
+		}
+		else {
+			$db_block = $r->fetch();
+			$current_hash = $db_block['block_hash'];
+		}
 	}
 	
 	do {
